requirements.txt
# Core dependencies
numpy
opencv-python
pillow
tqdm

# Face detection and analysis
insightface>=0.7.3
onnx
onnxruntime  # Use 'onnxruntime-gpu' if running on GPU
tensorflow>=2.13.0
torch  # Required internally by some InsightFace models

# Image quality scoring
imquality

# Video processing
ffmpeg-python
moviepy

# YouTube video downloading (if needed)
yt-dlp

# Optional: Google API support
google-api-python-client
google-auth
google-auth-oauthlib

# Utilities
scipy
psutil
